+++ internal/pg_documentdb_distributed/src/distribution/distributed_hooks.c
@@ -687,6 +687,47 @@ GetDistributedShardIndexOidsCore(uint64_t collectionId, int indexId, bool ignore
 }
 
 
+static const char *
+GetDistributedOperationCancellationQuery(int64 shardId, StringView *opIdView,
+										 int *nargs, Oid **argTypes, Datum **argValues,
+										 char **argNulls)
+{
+	StringInfo query = makeStringInfo();
+
+	/*
+	 * KillOp query attempts to cancel any operation that is still active but is a no-op
+	 * when the operation is already finished the connection state is 'idle', in order to
+	 * kill idle connection we have to force terminate the backend.
+	 *
+	 * For distributed cases we use citus overrides to cancel operations that are identified the
+	 * gpid
+	 */
+	appendStringInfo(query,
+					 " SELECT "
+					 "  CASE WHEN state = 'idle' THEN pg_terminate_backend($1)"
+					 "       ELSE pg_cancel_backend($1)"
+					 "  END "
+					 " FROM citus_stat_activity WHERE global_pid = $1 "
+					 " AND (EXTRACT(epoch FROM query_start) * 1000000)::numeric(20,0) = $2::numeric(20,0) "
+					 " AND NOT is_worker_query "
+					 "LIMIT 1");
+
+	*nargs = 2;
+	*argTypes = palloc(sizeof(Oid) * (*nargs));
+	*argValues = palloc(sizeof(Datum) * (*nargs));
+	*argNulls = palloc0(sizeof(char) * (*nargs));
+	(*argTypes)[0] = INT8OID;
+	(*argValues)[0] = Int64GetDatum(shardId);
+	(*argTypes)[1] = TEXTOID;
+	(*argValues)[1] = CStringGetTextDatum(opIdView->string);
+
+	(*argNulls)[0] = ' ';
+	(*argNulls)[1] = ' ';
+
+	return query->data;
+}
+
+
 /*
  * Register hook overrides for DocumentDB.
  */
@@ -725,6 +766,7 @@ InitializeDocumentDBDistributedHooks(void)
 	get_shard_index_oids_hook = GetDistributedShardIndexOidsCore;
 
 	update_postgres_index_hook = UpdateDistributedPostgresIndex;
+	get_operation_cancellation_query_hook = GetDistributedOperationCancellationQuery;
 
 	DistributedOperationsQuery =
 		"SELECT * FROM pg_stat_activity LEFT JOIN pg_catalog.get_all_active_transactions() ON process_id = pid JOIN pg_catalog.pg_dist_local_group ON TRUE";