+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### documentdb v0.109-0 (Unreleased) ###
 * Support collation with find positional queries *[Feature]*
+* Support for `killOp` administrative command *[Feature]*
 
 ### documentdb v0.108-0 (Unreleased) ###
 * Top-level `let` variables and `$$NOW` supported by default.